+++ GEOSwift/GEOSwiftError.swift
@@ -9,4 +9,5 @@ public enum GEOSwiftError: Error, Equatable {
     case invalidFeatureId
     case lengthIsZero
     case unexpectedEnvelopeResult(Geometry)
+    case negativeBufferWidth
 }